WIP Implement `OrderedStream` for `PropertyStream`
This seems like it could be useful. The behavior of an ordered stream
may bit a little odd in what it does and does not guarantee with the
behavior of dropping intermediate values that `PropertyStream` has, but
I believe this can still be useful. This looks like it also applies to
`NameOwnerChangedStream`, which is already an `OwnedStream`.

I think this approach should work, but there are a few details that
still need fixing, and I need to test it. But it seems good to create a
PR rather than just leave this sitting in my git stash.